Rename pages-interface to lazy-store
