+++ src/lazy-store.js
